Refactored the bot message delete feature & added trigger typing
- main.py:
    - Changed the bot message delete trigger from trash emote
      reaction to replying to the message with `del`. This
      fixes the issue where the function remains awaiting for
      the reaction, stopping the bot from processing more than
      1 search query.
    - Added a `trigger_typing` which shows that the bot is
      typing the response.

- metadata.py: Moved the footer message to a new field.

+++ main.py
@@ -1,5 +1,6 @@
 import os
 import re
+import asyncio
 
 import discord
 from discord.ext import commands
@@ -43,12 +44,10 @@ async def on_message(message):
     query = message.content.lower().splitlines()
     query = ' '.join(query)
 
-    def check(reaction, user):
-        return str(reaction.emoji) == '🗑️' and not user.bot and reaction.message.id == msg.id
-
     if str(message.channel.id) in channels:
 
         if re.search(r"^ao3\b", query) is not None:
+            await message.channel.trigger_typing()
             msg = query.replace("ao3", "")
             msg = msg.replace("ffn", "")
             embed_pg = ao3_metadata(msg)
@@ -57,12 +56,9 @@ async def on_message(message):
                 embed_pg = ffn_metadata(msg)
 
             msg = await message.channel.send(embed=embed_pg)
-            await msg.add_reaction("🗑️")
-
-            reaction, user = await client.wait_for('reaction_add', check=check)
-            await msg.delete()
 
         elif re.search(r"^ffn\b", query) is not None:
+            await message.channel.trigger_typing()
             msg = query.replace("ffn", "")
             msg = msg.replace("ao3", "")
             embed_pg = ffn_metadata(msg)
@@ -71,19 +67,15 @@ async def on_message(message):
                 embed_pg = ao3_metadata(msg)
 
             msg = await message.channel.send(embed=embed_pg)
-            await msg.add_reaction("🗑️")
-
-            reaction, user = await client.wait_for('reaction_add',  check=check)
-            await msg.delete()
 
         # if in code blocks
         elif re.search(r"`(.*?)`", query) is not None:
 
-            url_found = re.findall(
-                r"`(.*?)`", query.lower(), re.MULTILINE)
+            url_found = re.findall(r"`(.*?)`", query.lower(), re.MULTILINE)
+            url_found = url_found[:2]  # to limit the url to 2 only
 
-            url_found = url_found[:1]  # to limit the url to 1 only
             for i in url_found:
+                await message.channel.trigger_typing()
                 embed_pg = ffn_metadata(i)
 
                 if embed_pg is None:  # if not found in ffn, search in ao3
@@ -91,31 +83,38 @@ async def on_message(message):
                     embed_pg = ao3_metadata(msg2)
 
                 msg = await message.channel.send(embed=embed_pg)
-                await msg.add_reaction("🗑️")
-
-                reaction, user = await client.wait_for('reaction_add', check=check)
-                await msg.delete()
 
         elif re.search(URL_VALIDATE, query) is not None:
+
             url_found = re.findall(URL_VALIDATE, query.lower(), re.MULTILINE)
-            url_found = url_found[:1]  # to limit the url to 1 only
+            url_found = url_found[:2]  # to limit the url to 2 only
 
             for i in url_found:
+                await message.channel.trigger_typing()
                 if re.search(r"fanfiction.net\b",  i) is not None:
                     embed_pg = ffn_metadata(i)
                     msg = await message.channel.send(embed=embed_pg)
-                    await msg.add_reaction("🗑️")
-
-                    reaction, user = await client.wait_for('reaction_add', check=check)
-                    await msg.delete()
 
                 if re.search(r"archiveofourown.org\b", i) is not None:
                     embed_pg = ao3_metadata(i)
                     msg = await message.channel.send(embed=embed_pg)
-                    await msg.add_reaction("🗑️")
 
-                    reaction, user = await client.wait_for('reaction_add', check=check)
-                    await msg.delete()
+        elif re.search(r"^del\b", query) and message.reference is not None:
+            message_ref = message.reference  # message id of reply
+            message_to_delete = await message.channel.fetch_message(message_ref.message_id)
+
+            # if the messsage was created by a bot
+            if message_to_delete.author.bot:
+                await message_to_delete.delete()
+                await message.delete()
+            else:
+                await message.delete()
+                msg = await message.channel.send(
+                    embed=discord.Embed(
+                        description="I am not allowed to do that. I can only delete my messages."))
+                await asyncio.sleep(3)
+                await msg.delete()
+
 
 start_server()
 client.load_extension("cogs.settings")
